Move crate tarball related code into dedicated crate
