+++ Cargo.toml
@@ -38,6 +38,7 @@ base64 = "=0.13.1"
 cargo-registry-index = { path = "cargo-registry-index" }
 cargo-registry-markdown = { path = "cargo-registry-markdown" }
 cargo-registry-s3 = { path = "cargo-registry-s3" }
+cargo-registry-tarball = { path = "cargo-registry-tarball" }
 chrono = { version = "=0.4.26", default-features = false, features = ["serde"] }
 clap = { version = "=4.3.0", features = ["derive", "env", "unicode", "wrap_help"] }
 cookie = { version = "=0.17.0", features = ["secure"] }
