+++ cargo-registry-tarball/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "cargo-registry-tarball"
+version = "0.0.0"
+license = "MIT OR Apache-2.0"
+edition = "2021"
+
+[dependencies]
+derive_deref = "=1.1.1"
+flate2 = "=1.0.26"
+semver = { version = "=1.0.17", features = ["serde"] }
+serde = { version = "=1.0.163", features = ["derive"] }
+serde_json = "=1.0.96"
+tar = "=0.4.38"
+thiserror = "=1.0.40"
+toml = "=0.7.4"
+tracing = "=0.1.37"
+
+[dev-dependencies]
+claims = "=0.7.1"
