+++ cargo-registry-tarball/src/lib.rs
@@ -1,0 +1,192 @@
+#[cfg(test)]
+#[macro_use]
+extern crate claims;
+
+use crate::limit_reader::LimitErrorReader;
+pub use crate::manifest::Manifest;
+pub use crate::vcs_info::CargoVcsInfo;
+use flate2::read::GzDecoder;
+use std::io::Read;
+use std::path::Path;
+use tracing::instrument;
+
+mod limit_reader;
+mod manifest;
+mod vcs_info;
+
+#[derive(Debug)]
+pub struct TarballInfo {
+    pub manifest: Option<Manifest>,
+    pub vcs_info: Option<CargoVcsInfo>,
+}
+
+#[derive(Debug, thiserror::Error)]
+pub enum TarballError {
+    #[error("uploaded tarball is malformed or too large when decompressed")]
+    Malformed(#[source] std::io::Error),
+    #[error("invalid tarball uploaded")]
+    Invalid,
+    #[error(transparent)]
+    IO(#[from] std::io::Error),
+}
+
+#[instrument(skip_all, fields(%pkg_name))]
+pub fn verify_tarball(
+    pkg_name: &str,
+    tarball: &[u8],
+    max_unpack: u64,
+) -> Result<TarballInfo, TarballError> {
+    // All our data is currently encoded with gzip
+    let decoder = GzDecoder::new(tarball);
+
+    // Don't let gzip decompression go into the weeeds, apply a fixed cap after
+    // which point we say the decompressed source is "too large".
+    let decoder = LimitErrorReader::new(decoder, max_unpack);
+
+    // Use this I/O object now to take a peek inside
+    let mut archive = tar::Archive::new(decoder);
+
+    let vcs_info_path = Path::new(&pkg_name).join(".cargo_vcs_info.json");
+    let mut vcs_info = None;
+
+    let manifest_path = Path::new(&pkg_name).join("Cargo.toml");
+    let mut manifest = None;
+
+    for entry in archive.entries()? {
+        let mut entry = entry.map_err(TarballError::Malformed)?;
+
+        // Verify that all entries actually start with `$name-$vers/`.
+        // Historically Cargo didn't verify this on extraction so you could
+        // upload a tarball that contains both `foo-0.1.0/` source code as well
+        // as `bar-0.1.0/` source code, and this could overwrite other crates in
+        // the registry!
+        let entry_path = entry.path()?;
+        if !entry_path.starts_with(pkg_name) {
+            return Err(TarballError::Invalid);
+        }
+        if entry_path == vcs_info_path {
+            let mut contents = String::new();
+            entry.read_to_string(&mut contents)?;
+            vcs_info = CargoVcsInfo::from_contents(&contents).ok();
+        } else if entry_path == manifest_path {
+            // Try to extract and read the Cargo.toml from the tarball, silently
+            // erroring if it cannot be read.
+            let mut contents = String::new();
+            entry.read_to_string(&mut contents)?;
+            manifest = toml::from_str(&contents).ok();
+        }
+
+        // Historical versions of the `tar` crate which Cargo uses internally
+        // don't properly prevent hard links and symlinks from overwriting
+        // arbitrary files on the filesystem. As a bit of a hammer we reject any
+        // tarball with these sorts of links. Cargo doesn't currently ever
+        // generate a tarball with these file types so this should work for now.
+        let entry_type = entry.header().entry_type();
+        if entry_type.is_hard_link() || entry_type.is_symlink() {
+            return Err(TarballError::Invalid);
+        }
+    }
+
+    Ok(TarballInfo { manifest, vcs_info })
+}
+
+#[cfg(test)]
+mod tests {
+    use super::verify_tarball;
+    use flate2::read::GzEncoder;
+    use std::io::{Read, Write};
+
+    fn add_file<W: Write>(pkg: &mut tar::Builder<W>, path: &str, content: &[u8]) {
+        let mut header = tar::Header::new_gnu();
+        header.set_size(content.len() as u64);
+        header.set_cksum();
+        pkg.append_data(&mut header, path, content).unwrap();
+    }
+
+    #[test]
+    fn verify_tarball_test() {
+        let mut pkg = tar::Builder::new(vec![]);
+        add_file(&mut pkg, "foo-0.0.1/Cargo.toml", b"");
+        let mut serialized_archive = vec![];
+        GzEncoder::new(pkg.into_inner().unwrap().as_slice(), Default::default())
+            .read_to_end(&mut serialized_archive)
+            .unwrap();
+
+        let limit = 512 * 1024 * 1024;
+        assert_eq!(
+            verify_tarball("foo-0.0.1", &serialized_archive, limit)
+                .unwrap()
+                .vcs_info,
+            None
+        );
+        assert_err!(verify_tarball("bar-0.0.1", &serialized_archive, limit));
+    }
+
+    #[test]
+    fn verify_tarball_test_incomplete_vcs_info() {
+        let mut pkg = tar::Builder::new(vec![]);
+        add_file(&mut pkg, "foo-0.0.1/Cargo.toml", b"");
+        add_file(
+            &mut pkg,
+            "foo-0.0.1/.cargo_vcs_info.json",
+            br#"{"unknown": "field"}"#,
+        );
+        let mut serialized_archive = vec![];
+        GzEncoder::new(pkg.into_inner().unwrap().as_slice(), Default::default())
+            .read_to_end(&mut serialized_archive)
+            .unwrap();
+        let limit = 512 * 1024 * 1024;
+        let vcs_info = verify_tarball("foo-0.0.1", &serialized_archive, limit)
+            .unwrap()
+            .vcs_info
+            .unwrap();
+        assert_eq!(vcs_info.path_in_vcs, "");
+    }
+
+    #[test]
+    fn verify_tarball_test_vcs_info() {
+        let mut pkg = tar::Builder::new(vec![]);
+        add_file(&mut pkg, "foo-0.0.1/Cargo.toml", b"");
+        add_file(
+            &mut pkg,
+            "foo-0.0.1/.cargo_vcs_info.json",
+            br#"{"path_in_vcs": "path/in/vcs"}"#,
+        );
+        let mut serialized_archive = vec![];
+        GzEncoder::new(pkg.into_inner().unwrap().as_slice(), Default::default())
+            .read_to_end(&mut serialized_archive)
+            .unwrap();
+        let limit = 512 * 1024 * 1024;
+        let vcs_info = verify_tarball("foo-0.0.1", &serialized_archive, limit)
+            .unwrap()
+            .vcs_info
+            .unwrap();
+        assert_eq!(vcs_info.path_in_vcs, "path/in/vcs");
+    }
+
+    #[test]
+    fn verify_tarball_test_manifest() {
+        let mut pkg = tar::Builder::new(vec![]);
+        add_file(
+            &mut pkg,
+            "foo-0.0.1/Cargo.toml",
+            br#"
+[package]
+rust-version = "1.59"
+readme = "README.md"
+repository = "https://github.com/foo/bar"
+"#,
+        );
+        let mut serialized_archive = vec![];
+        GzEncoder::new(pkg.into_inner().unwrap().as_slice(), Default::default())
+            .read_to_end(&mut serialized_archive)
+            .unwrap();
+
+        let limit = 512 * 1024 * 1024;
+        let tarball_info = assert_ok!(verify_tarball("foo-0.0.1", &serialized_archive, limit));
+        let manifest = assert_some!(tarball_info.manifest);
+        assert_some_eq!(manifest.package.readme, "README.md");
+        assert_some_eq!(manifest.package.repository, "https://github.com/foo/bar");
+        assert_some_eq!(manifest.package.rust_version, "1.59");
+    }
+}
